+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/implementation/http/ReactorNettyClient.java
@@ -149,13 +149,13 @@ public class ReactorNettyClient implements HttpClient {
                         .maxChunkSize(this.httpClientConfig.getMaxChunkSize())
                         .validateHeaders(true));
 
-        if (httpClientConfig.getHttp2Config().isEnabled()) {
+        if (httpClientConfig.getHttp2ConnectionConfig().isEnabled()) {
             this.httpClient = this.httpClient
                 .secure(sslContextSpec ->
                     sslContextSpec.sslContext(
                         configs.getSslContext(
                             httpClientConfig.isServerCertValidationDisabled(),
-                            httpClientConfig.getHttp2Config().isEnabled()
+                            true
                         )))
                 .protocol(HttpProtocol.H2, HttpProtocol.HTTP11)
                 .doOnConnected((connection -> {